remove BreakMethodArgumentsFixer, interated in LineLengthFixer, also move test
